# backend/app/docs/swagger.yaml
basePath: /api/v1
definitions:
  types.BriefComment:
    description: 简要的评论信息展示
    properties:
      active:
        example: active
        type: string
      avatar_url:
        example: https://example.com/avatar.jpg
        type: string
      comment_count:
        example: 3
        type: integer
      content:
        example: 这是一条评论内容
        type: string
      create_at:
        example: "2025-04-27T16:10:08.5Z"
        type: string
      id:
        example: 456
        type: integer
      like_count:
        example: 10
        type: integer
      post_id:
        example: 123
        type: integer
      poster_id:
        example: 789
        type: integer
      poster_nickname:
        example: 张三
        type: string
    type: object
...
  types.BriefSubComment:
    description: 简要的子评论信息展示
    properties:
      active:
        example: 30分钟前
        type: string
      avatar_url:
        example: https://example.com/avatar.jpg
        type: string
      content:
        example: 这是一条回复评论
        type: string
      create_at:
        example: "2025-04-27T16:10:08.5Z"
        type: string
      id:
        example: 789
        type: integer
      like_count:
        example: 5
        type: integer
      parent:
        example: 456
        type: integer
      post_id:
        example: 123
        type: integer
      poster_id:
        example: 456
        type: integer
      poster_nickname:
        example: 李四
        type: string
      reply:
        example: 101
        type: integer
    type: object
...
  types.CommentCreateReq:
    description: 创建评论的请求参数
    properties:
      content:
        example: 这是一条评论内容
        maxLength: 500
        minLength: 1
        type: string
      post_id:
        example: 123
        type: integer
    required:
    - content
    - post_id
    type: object
  types.CommentCreateResp:
    description: 创建评论成功后返回的响应数据
    properties:
      id:
        example: 456
        type: integer
    type: object
  types.CommentLikeReq:
    description: 评论点赞/取消点赞的请求参数
    properties:
      comment_id:
        example: 456
        type: integer
      like_status:
        enum:
        - 0
        - 1
        example: 1
        type: integer
    required:
    - comment_id
    - like_status
    type: object
  types.CommentLikeResp:
    description: 评论点赞操作后返回的响应数据
    properties:
      like_count:
        example: 10
        type: integer
      like_status:
        example: 1
        type: integer
    type: object
  types.CommentListReq:
    description: 获取评论列表的请求参数
    properties:
      last_cid:
        example: 100
        type: integer
      post_id:
        example: 123
        type: integer
      size:
        example: 10
        maximum: 50
        minimum: 1
        type: integer
    required:
    - post_id
    - size
    type: object
  types.CreatePostResp:
    description: 创建帖子成功后返回的响应数据
    properties:
      id:
        example: 123
        type: integer
    type: object
...
  types.PostResp:
    description: 帖子详情的响应数据
    properties:
      active:
        enum:
        - active
        - locked
        - disabled
        type: string
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      collect_count:
        type: integer
      comment_count:
        type: integer
      content:
        type: string
      create_at:
        type: string
      id:
        type: integer
      image_urls:
        example:
        - https://example.com/image1.jpg
        - https://example.com/image2.jpg
        items:
          type: string
        type: array
      like_count:
        type: integer
      like_status:
        example: 1
        type: integer
      poster_campus:
        example: 翡翠湖校区
        type: string
      poster_department:
        example: 计算机学院
        type: string
      poster_grade:
        example: "2022"
        type: string
      poster_id:
        type: integer
      poster_nickname:
        example: 张三
        type: string
      title:
        type: string
      update_at:
        type: string
      view_count:
        type: integer
    required:
    - active
    type: object
  types.SubCommentCreateReq:
    description: 创建子评论(回复评论)的请求参数
    properties:
      content:
        example: 这是一条回复评论
        maxLength: 500
        minLength: 1
        type: string
      parent:
        example: 456
        type: integer
      post_id:
        example: 123
        type: integer
      reply:
        example: 789
        type: integer
    required:
    - content
    - parent
    - post_id
    type: object
  types.SubCommentCreateResp:
    description: 创建子评论成功后返回的响应数据
    properties:
      id:
        example: 789
        type: integer
    type: object
  types.SubCommentLikeReq:
    description: 子评论点赞/取消点赞的请求参数
    properties:
      like_status:
        enum:
        - 0
        - 1
        example: 1
        type: integer
      subcomment_id:
        example: 789
        type: integer
    required:
    - like_status
    - subcomment_id
    type: object
  types.SubCommentLikeResp:
    description: 子评论点赞操作后返回的响应数据
    properties:
      like_count:
        example: 5
        type: integer
      like_status:
        example: 1
        type: integer
    type: object
  types.SubCommentListReq:
    description: 获取子评论列表的请求参数
    properties:
      last_scid:
        example: 100
        type: integer
      parent:
        example: 456
        type: integer
      size:
        example: 10
        maximum: 50
        minimum: 1
        type: integer
    required:
    - parent
    - size
    type: object
  types.UpdateUserInfoReq:
    description: 更新用户信息的请求参数
    properties:
      account_status:
        enum:
        - active
        - locked
        - disabled
        example: active
        type: string
      nickname:
        example: 李华2333
        maxLength: 50
        minLength: 1
        type: string
      signature:
        example: 这个人说了什么
        type: string
      tags:
        example:
        - 唱
        - 跳
        - rap
        - 篮球
        items:
          type: string
        type: array
    required:
    - account_status
    - nickname
    type: object
  types.UpdateUserInfoResp:
    description: 更新用户信息后返回的响应数据
    properties:
      account_status:
        enum:
        - active
        - locked
        - disabled
        example: active
        type: string
      avatar_url:
        example: https://127.0.0.1:8080/api/v1/user/auth/avatar/1
        type: string
      backimage_url:
        example: https://127.0.0.1:8080/api/v1/user/auth/backimage/1
        type: string
      campus:
        example: 翡翠湖校区
        type: string
      class:
        example: 信息计22-2班
        type: string
      create_at:
        example: "2025-04-27T16:10:08.5Z"
        type: string
      cultivate_type:
        example: 主修
        type: string
      department:
        example: 数学学院
        type: string
      enrollment_date:
        example: "2022-09-01"
        type: string
      grade:
        example: "2022"
        type: string
      graduate_date:
        example: "2026-07-01"
        type: string
      id:
        example: 1
        type: integer
      level:
        example: 本科
        type: string
      major:
        example: 信息与计算科学
        type: string
      nickname:
        example: 傅益忠
        maxLength: 50
        minLength: 1
        type: string
      sex:
        example: 男
        type: string
      signature:
        example: 这人啥也没说
        type: string
      tags:
        description: 数组类型用逗号分隔示例
        example:
        - 学习
        - 运动
        - 音乐
        items:
          type: string
        type: array
      username:
        example: "2022210826"
        maxLength: 50
        minLength: 3
        type: string
      username_zh:
        example: 傅益忠
        type: string
    required:
    - account_status
    - nickname
    - username
    type: object
  types.UploadImageResp:
    description: 上传图片成功后返回的响应数据
    properties:
      url:
        example: https://127.0.0.1:8080/api/v1/user/auth/avatar/1
        type: string
    type: object
info:
  contact: {}
  title: 接口文档
  version: "1.0"
paths:
  /comment/auth/createreply:
    post:
      consumes:
      - application/json
      description: 输入正文和父评论id和回复用户id创建一条回复并返回这个回复的id
      parameters:
      - description: 请求体
        in: body
        name: edit
        required: true
        schema:
          $ref: '#/definitions/types.SubCommentCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/types.SubCommentCreateResp'
      security:
      - BearerAuth: []
      summary: 创建回复
      tags:
      - Comment模块
  /comment/auth/like:
    post:
      consumes:
      - application/json
      description: 还没做
      parameters:
      - description: 请求体
        in: body
        name: edit
        required: true
        schema:
          $ref: '#/definitions/types.CommentLikeReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/types.CommentLikeResp'
      security:
      - BearerAuth: []
      summary: 评论点赞【TODO】
      tags:
      - Comment模块
  /comment/auth/likereply:
    post:
      consumes:
      - application/json
      description: 还没做
      parameters:
      - description: 请求体
        in: body
        name: edit
        required: true
        schema:
          $ref: '#/definitions/types.SubCommentLikeReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/types.SubCommentLikeResp'
      security:
      - BearerAuth: []
      summary: 回复点赞【TODO】
      tags:
      - Comment模块
  /comment/auth/list:
    post:
      consumes:
      - application/json
      description: 输入帖子id返回评论列表
      parameters:
      - description: 请求体
        in: body
        name: edit
        required: true
        schema:
          $ref: '#/definitions/types.CommentListReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            description: 评论列表的响应数据
            items:
              $ref: '#/definitions/types.BriefComment'
            type: array
      security:
      - BearerAuth: []
      summary: 评论列表
      tags:
      - Comment模块
  /comment/auth/listreply:
    post:
      consumes:
      - application/json
      description: 输入父评论id返回回复列表
      parameters:
      - description: 请求体
        in: body
        name: edit
        required: true
        schema:
          $ref: '#/definitions/types.SubCommentListReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            description: 子评论列表的响应数据
            items:
              $ref: '#/definitions/types.BriefSubComment'
            type: array
      security:
      - BearerAuth: []
      summary: 回复列表
      tags:
      - Comment模块
  /commmnt/auth/create:
    post:
      consumes:
      - application/json
      description: 输入正文和帖子id创建评论并返回评论id
      parameters:
      - description: 请求体
        in: body
        name: edit
        required: true
        schema:
          $ref: '#/definitions/types.CommentCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/types.CommentCreateResp'
      security:
      - BearerAuth: []
      summary: 创建评论
      tags:
      - Comment模块
  /post/auth/{id}:
    get:
      consumes:
      - application/json
      description: 通过帖子id直接查看帖子
      parameters:
      - description: 帖子ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/types.PostResp'
      security:
      - BearerAuth: []
      summary: 帖子查看
      tags:
      - Post模块
  /post/auth/create:
    post:
      consumes:
      - multipart/form-data
      description: 输入标题和正文创建帖子，并返回新帖子id
      parameters:
      - description: 背景图文件
        in: formData
        name: postimage
        required: true
        type: file
      - description: JSON请求数据
        in: formData
        name: json
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/types.CreatePostResp'
      security:
      - BearerAuth: []
      summary: 创建帖子
      tags:
      - Post模块
...
  /post/auth/postimage/{id}:
    get:
      consumes:
      - application/json
      description: 通过帖子id直接查看帖子
      parameters:
      - description: 帖子图片ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 帖子图片文件
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: 帖子图片查看
      tags:
      - Post模块
...
  /user/auth/avatar/{id}:
    get:
      description: 根据用户ID返回头像数据
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 头像图片文件
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: 获取用户头像
      tags:
      - User模块
  /user/auth/backimage/{id}:
    get:
      description: 根据用户ID返回背景图数据
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 背景图片文件
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: 获取用户背景图
      tags:
      - User模块
...
  /user/auth/update_userinfo:
    post:
      consumes:
      - application/json
      description: 更改用户信息
      parameters:
      - description: 请求体
        in: body
        name: update_userinfo
        required: true
        schema:
          $ref: '#/definitions/types.UpdateUserInfoReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/types.UpdateUserInfoResp'
      security:
      - BearerAuth: []
      summary: 更改用户信息
      tags:
      - User模块
  /user/auth/upload_avatar:
    post:
      consumes:
      - multipart/form-data
      description: 通过form-data上传头像
      parameters:
      - description: 图片文件
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/types.UploadImageResp'
      security:
      - BearerAuth: []
      summary: 更新头像
      tags:
      - User模块
  /user/auth/upload_backimage:
    post:
      consumes:
      - multipart/form-data
      description: 通过form-data上传背景图
      parameters:
      - description: 背景图文件
        in: formData
        name: backimage
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/types.UploadImageResp'
      security:
      - BearerAuth: []
      summary: 更新背景图
      tags:
      - User模块
...
securityDefinitions:
  BearerAuth:
    description: Bearer Token 格式：`Bearer {token}`（注意空格）
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"